+++ src/components/VideoProvider/useLocalTracks/useLocalTracks.ts
@@ -1,5 +1,12 @@
+/**
+ * #ANGLES_EDIT
+ *
+ * Now creates a LocalDataTrack for the local participant in addition to audio and video tracks.
+ * The created LocalDataTrack is included in the `localTracks` property returned.
+ */
+
 import { useCallback, useEffect, useState } from 'react';
-import Video, { LocalVideoTrack, LocalAudioTrack } from 'twilio-video';
+import Video, { LocalVideoTrack, LocalAudioTrack, LocalDataTrack } from 'twilio-video';
 
 export function useLocalAudioTrack() {
   const [track, setTrack] = useState<LocalAudioTrack>();
@@ -19,6 +26,20 @@ export function useLocalAudioTrack() {
       };
     }
   }, [track]);
+
+  return track;
+}
+
+export function useLocalDataTrack() {
+  const [track, setTrack] = useState<LocalDataTrack>();
+
+  useEffect(() => {
+    // Not sure if this is the right way to create a data track?
+    // The docs had some weird Promise stuff going on. https://www.twilio.com/docs/video/using-datatrack-api
+    setTrack(new LocalDataTrack());
+  }, []);
+
+  // LocalDataTracks don't seem to emit any events. So no need to handle stopping, etc...
 
   return track;
 }
@@ -61,10 +82,12 @@ export function useLocalVideoTrack() {
 export default function useLocalTracks() {
   const audioTrack = useLocalAudioTrack();
   const [videoTrack, getLocalVideoTrack] = useLocalVideoTrack();
+  const dataTrack = useLocalDataTrack();
 
-  const localTracks = [audioTrack, videoTrack].filter(track => track !== undefined) as (
+  const localTracks = [audioTrack, videoTrack, dataTrack].filter(track => track !== undefined) as (
     | LocalAudioTrack
     | LocalVideoTrack
+    | LocalDataTrack
   )[];
 
   return { localTracks, getLocalVideoTrack };